added Prover trait
