+++ examples/src/fibonacci/fib8/air.rs
@@ -4,17 +4,15 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
+use super::{BaseElement, FieldElement, TRACE_WIDTH};
 use crate::utils::are_equal;
 use winterfell::{
-    math::{fields::f128::BaseElement, FieldElement},
-    Air, AirContext, Assertion, EvaluationFrame, ExecutionTrace, ProofOptions, TraceInfo,
+    Air, AirContext, Assertion, EvaluationFrame, ProofOptions, TraceInfo,
     TransitionConstraintDegree,
 };
 
 // FIBONACCI AIR
 // ================================================================================================
-
-const TRACE_WIDTH: usize = 2;
 
 pub struct Fib8Air {
     context: AirContext<BaseElement>,
@@ -79,43 +77,4 @@ impl Air for Fib8Air {
             Assertion::single(1, last_step, self.result),
         ]
     }
-}
-
-// FIBONACCI TRACE BUILDER
-// ================================================================================================
-
-pub fn build_trace(length: usize) -> ExecutionTrace<BaseElement> {
-    assert!(
-        length.is_power_of_two(),
-        "sequence length must be a power of 2"
-    );
-
-    // initialize the trace with 7th and 8th terms of Fibonacci sequence (skipping the first 6)
-    let n0 = BaseElement::ONE;
-    let n1 = BaseElement::ONE;
-    let n2 = n0 + n1;
-    let n3 = n1 + n2;
-    let n4 = n2 + n3;
-    let n5 = n3 + n4;
-    let n6 = n4 + n5;
-    let n7 = n5 + n6;
-
-    let mut reg0 = vec![n6];
-    let mut reg1 = vec![n7];
-
-    for i in 0..(length / 8 - 1) {
-        let n0 = reg0[i] + reg1[i];
-        let n1 = reg1[i] + n0;
-        let n2 = n0 + n1;
-        let n3 = n1 + n2;
-        let n4 = n2 + n3;
-        let n5 = n3 + n4;
-        let n6 = n4 + n5;
-        let n7 = n5 + n6;
-
-        reg0.push(n6);
-        reg1.push(n7);
-    }
-
-    ExecutionTrace::init(vec![reg0, reg1])
 }
