+++ examples/src/fibonacci/fib8/mod.rs
@@ -10,14 +10,22 @@ use log::debug;
 use std::time::Instant;
 use winterfell::{
     math::{fields::f128::BaseElement, log2, FieldElement},
-    ProofOptions, StarkProof, VerifierError,
+    ExecutionTrace, ProofOptions, Prover, StarkProof, VerifierError,
 };
 
 mod air;
-use air::{build_trace, Fib8Air};
+use air::Fib8Air;
+
+mod prover;
+use prover::Fib8Prover;
 
 #[cfg(test)]
 mod tests;
+
+// CONSTANTS
+// ================================================================================================
+
+const TRACE_WIDTH: usize = 2;
 
 // FIBONACCI EXAMPLE
 // ================================================================================================
@@ -70,9 +78,12 @@ impl Example for Fib8Example {
             self.sequence_length
         );
 
+        // create a prover
+        let prover = Fib8Prover::new(self.options.clone());
+
         // generate execution trace
         let now = Instant::now();
-        let trace = build_trace(self.sequence_length);
+        let trace = prover.build_trace(self.sequence_length);
         let trace_width = trace.width();
         let trace_length = trace.length();
         debug!(
@@ -83,7 +94,7 @@ impl Example for Fib8Example {
         );
 
         // generate the proof
-        winterfell::prove::<Fib8Air>(trace, self.result, self.options.clone()).unwrap()
+        prover.prove(trace, self.result).unwrap()
     }
 
     fn verify(&self, proof: StarkProof) -> Result<(), VerifierError> {
