+++ examples/src/fibonacci/mulfib2/air.rs
@@ -7,7 +7,7 @@
 use crate::utils::are_equal;
 use winterfell::{
     math::{fields::f128::BaseElement, FieldElement},
-    Air, AirContext, Assertion, EvaluationFrame, ExecutionTrace, ProofOptions, TraceInfo,
+    Air, AirContext, Assertion, EvaluationFrame, ProofOptions, TraceInfo,
     TransitionConstraintDegree,
 };
 
@@ -72,24 +72,4 @@ impl Air for MulFib2Air {
             Assertion::single(0, last_step, self.result),
         ]
     }
-}
-
-// FIBONACCI TRACE BUILDER
-// ================================================================================================
-
-pub fn build_trace(length: usize) -> ExecutionTrace<BaseElement> {
-    assert!(
-        length.is_power_of_two(),
-        "sequence length must be a power of 2"
-    );
-
-    let mut reg0 = vec![BaseElement::new(1)];
-    let mut reg1 = vec![BaseElement::new(2)];
-
-    for i in 0..(length / 2 - 1) {
-        reg0.push(reg0[i] * reg1[i]);
-        reg1.push(reg1[i] * reg0[i + 1]);
-    }
-
-    ExecutionTrace::init(vec![reg0, reg1])
 }
