+++ examples/src/fibonacci/mulfib2/mod.rs
@@ -10,11 +10,14 @@ use log::debug;
 use std::time::Instant;
 use winterfell::{
     math::{fields::f128::BaseElement, log2, FieldElement},
-    ProofOptions, StarkProof, VerifierError,
+    ExecutionTrace, ProofOptions, Prover, StarkProof, VerifierError,
 };
 
 mod air;
-use air::{build_trace, MulFib2Air};
+use air::MulFib2Air;
+
+mod prover;
+use prover::MulFib2Prover;
 
 #[cfg(test)]
 mod tests;
@@ -70,9 +73,12 @@ impl Example for MulFib2Example {
             sequence_length
         );
 
+        // create a prover
+        let prover = MulFib2Prover::new(self.options.clone());
+
         // generate execution trace
         let now = Instant::now();
-        let trace = build_trace(sequence_length);
+        let trace = prover.build_trace(sequence_length);
         let trace_width = trace.width();
         let trace_length = trace.length();
         debug!(
@@ -83,7 +89,7 @@ impl Example for MulFib2Example {
         );
 
         // generate the proof
-        winterfell::prove::<MulFib2Air>(trace, self.result, self.options.clone()).unwrap()
+        prover.prove(trace, self.result).unwrap()
     }
 
     fn verify(&self, proof: StarkProof) -> Result<(), VerifierError> {
