+++ examples/src/merkle/mod.rs
@@ -4,6 +4,10 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
+use crate::utils::rescue::{
+    self, CYCLE_LENGTH as HASH_CYCLE_LEN, NUM_ROUNDS as NUM_HASH_ROUNDS,
+    STATE_WIDTH as HASH_STATE_WIDTH,
+};
 use crate::{
     utils::rescue::{Hash, Rescue128},
     Example, ExampleOptions,
@@ -13,15 +17,23 @@ use rand_utils::{rand_value, rand_vector};
 use std::time::Instant;
 use winterfell::{
     crypto::{Digest, MerkleTree},
-    math::{fields::f128::BaseElement, log2, StarkField},
-    ProofOptions, StarkProof, VerifierError,
+    math::{fields::f128::BaseElement, log2, FieldElement, StarkField},
+    ExecutionTrace, ProofOptions, Prover, StarkProof, VerifierError,
 };
 
 mod air;
-use air::{build_trace, MerkleAir, PublicInputs};
+use air::{MerkleAir, PublicInputs};
+
+mod prover;
+use prover::MerkleProver;
 
 #[cfg(test)]
 mod tests;
+
+// CONSTANTS
+// ================================================================================================
+
+const TRACE_WIDTH: usize = 7;
 
 // MERKLE AUTHENTICATION PATH EXAMPLE
 // ================================================================================================
@@ -90,8 +102,10 @@ impl Example for MerkleExample {
             ---------------------",
             self.path.len()
         );
+        let prover = MerkleProver::new(self.options.clone());
+
         let now = Instant::now();
-        let trace = build_trace(self.value, &self.path, self.index);
+        let trace = prover.build_trace(self.value, &self.path, self.index);
         let trace_length = trace.length();
         debug!(
             "Generated execution trace of {} registers and 2^{} steps in {} ms",
@@ -104,7 +118,7 @@ impl Example for MerkleExample {
         let pub_inputs = PublicInputs {
             tree_root: self.tree_root.to_elements(),
         };
-        winterfell::prove::<MerkleAir>(trace, pub_inputs, self.options.clone()).unwrap()
+        prover.prove(trace, pub_inputs).unwrap()
     }
 
     fn verify(&self, proof: StarkProof) -> Result<(), VerifierError> {
