+++ examples/src/rescue/rescue_128/mod.rs
@@ -9,17 +9,26 @@ use log::debug;
 use std::time::Instant;
 use winterfell::{
     math::{fields::f128::BaseElement, log2, FieldElement},
-    ProofOptions, StarkProof, VerifierError,
+    ExecutionTrace, ProofOptions, Prover, StarkProof, VerifierError,
 };
 
 #[allow(clippy::module_inception)]
 mod rescue;
 
 mod air;
-use air::{build_trace, PublicInputs, RescueAir};
+use air::{PublicInputs, RescueAir};
+
+mod prover;
+use prover::RescueProver;
 
 #[cfg(test)]
 mod tests;
+
+// CONSTANTS
+// ================================================================================================
+
+const CYCLE_LENGTH: usize = 16;
+const NUM_HASH_ROUNDS: usize = 14;
 
 // RESCUE HASH CHAIN EXAMPLE
 // ================================================================================================
@@ -75,8 +84,11 @@ impl Example for RescueExample {
             ---------------------",
             self.chain_length
         );
+
+        let prover = RescueProver::new(self.options.clone());
+
         let now = Instant::now();
-        let trace = build_trace(self.seed, self.chain_length);
+        let trace = prover.build_trace(self.seed, self.chain_length);
         let trace_length = trace.length();
         debug!(
             "Generated execution trace of {} registers and 2^{} steps in {} ms",
@@ -90,7 +102,7 @@ impl Example for RescueExample {
             seed: self.seed,
             result: self.result,
         };
-        winterfell::prove::<RescueAir>(trace, pub_inputs, self.options.clone()).unwrap()
+        prover.prove(trace, pub_inputs).unwrap()
     }
 
     fn verify(&self, proof: StarkProof) -> Result<(), VerifierError> {
