+++ examples/src/rescue/rescue_63/air.rs
@@ -8,14 +8,13 @@ use super::rescue;
 use crate::utils::{are_equal, is_zero, not, EvaluationResult};
 use winterfell::{
     math::{fields::f63::BaseElement, FieldElement},
-    Air, AirContext, Assertion, ByteWriter, EvaluationFrame, ExecutionTrace, ProofOptions,
-    Serializable, TraceInfo, TransitionConstraintDegree,
+    Air, AirContext, Assertion, ByteWriter, EvaluationFrame, ProofOptions, Serializable, TraceInfo,
+    TransitionConstraintDegree,
 };
 
 // CONSTANTS
 // ================================================================================================
 
-const NUM_HASH_ROUNDS: usize = 7;
 const CYCLE_LENGTH: usize = 8;
 const TRACE_WIDTH: usize = 14;
 
@@ -161,49 +160,4 @@ fn enforce_hash_copy<E: FieldElement>(result: &mut [E], current: &[E], next: &[E
     result.agg_constraint(11, flag, is_zero(next[11]));
     result.agg_constraint(12, flag, is_zero(next[12]));
     result.agg_constraint(13, flag, is_zero(next[13]));
-}
-
-// RESCUE TRACE GENERATOR
-// ================================================================================================
-
-pub fn build_trace(seed: [BaseElement; 7], iterations: usize) -> ExecutionTrace<BaseElement> {
-    // allocate memory to hold the trace table
-    let trace_length = iterations * CYCLE_LENGTH;
-    let mut trace = ExecutionTrace::new(14, trace_length);
-
-    trace.fill(
-        |state| {
-            // initialize first state of the computation
-            state[0] = seed[0];
-            state[1] = seed[1];
-            state[2] = seed[2];
-            state[3] = seed[3];
-            state[4] = seed[4];
-            state[5] = seed[5];
-            state[6] = seed[6];
-            state[7] = BaseElement::ZERO;
-            state[8] = BaseElement::ZERO;
-            state[9] = BaseElement::ZERO;
-            state[10] = BaseElement::ZERO;
-            state[11] = BaseElement::ZERO;
-            state[12] = BaseElement::ZERO;
-            state[13] = BaseElement::ZERO;
-        },
-        |step, state| {
-            // execute the transition function for all steps
-            if (step % CYCLE_LENGTH) < NUM_HASH_ROUNDS {
-                rescue::apply_round(state, step);
-            } else {
-                state[7] = BaseElement::ZERO;
-                state[8] = BaseElement::ZERO;
-                state[9] = BaseElement::ZERO;
-                state[10] = BaseElement::ZERO;
-                state[11] = BaseElement::ZERO;
-                state[12] = BaseElement::ZERO;
-                state[13] = BaseElement::ZERO;
-            }
-        },
-    );
-
-    trace
 }
